### 2025-04-17 #701 @marcburri
Presubmission reviser
Submitting Author Name: Marc Burri
Submitting Author Github Handle: @marcburri
Repository:  https://github.com/p-wegmueller/reviser
Submission type: Pre-submission

---

```
Package: reviser
Title: Tools for Studying Revision Properties in Real-Time Time Series Vintages
Version: 0.1.0.9000
    person("Marc", "Burri", ,"marc.burri91@gmail.com", role = c("aut", "cre", "cph"),
    person(given = "Philipp", family = "Wegmueller",
        email = "philipp.wemueller@seco.admin.ch", role = c("aut", "cph"))
  )
Description: Provides tools to analyze revision properties in real-time time series data.
Maintainers: c(
    person("Marc", "Burri", ,"marc.burri91@gmail.com",
    person(given = "Philipp", family = "Wegmueller",
        email = "philipp.wemueller@seco.admin.ch")
  )
    car,
    sandwich,
    systemfit,
    calculus,
    scales,
    glue,
    KFAS
    tsbox,
    purrr

```

The reviser package clearly falls under Time Series Analyses as it is purpose-built for handling real-time time series vintages, a core area of time series econometrics. It provides tools to calculate, visualize, and analyze temporal data revisions and patterns across different time-indexed data releases, which is fundamental for economic analyses  and revision impact assessment.

Yes, many of these standards

The target audience of the reviser package includes applied economists, statisticians, data scientists, and official statisticians—particularly those working with macroeconomic indicators such as GDP, inflation, or labor market statistics. It is especially useful for professionals dealing with real-time data, where the first release of a time series is subject to subsequent revisions.

Yes, [rjd3revisions](https://rjdverse.github.io/rjd3revisions/). The reviser package sets itself apart from rjd3revisions not only through its focus on advanced analysis of efficient releases and nowcasting performance, but also in its pure R implementation, which avoids external dependencies. In contrast, rjd3revisions relies heavily on Java via the JDemetra+ platform, which can make setup and integration more complex. reviser offers a lightweight, R-native solution for revision analysis, combining user-friendly tools for data wrangling, visualization, and evaluation of release efficiency.

No
