### 2025-04-18 #701 @maurolepore
@marcburri thanks a lot for sharing your work with rOpenSci.

I see it's proposed under a statistical-package category. I'll go ahead and ask the bot to perform the relevant checks.
