### 2025-04-21 #701 @mpadge
@marcburri Thanks for your submission. Your package definitely looks like a great fit for statistical software review. Can you please follow the procedures described in [the `srr` vignette](https://docs.ropensci.org/srr/articles/srr-stats.html) to insert `srrstats` roclet tags throughout your software? You can use the `srr_stats_pre_submit()` function to check compliance. Once you're done, let us know here, or call `@ropensci-review-bot check srr` like above to confirm. Thanks, and feel free to ask any further questions in the meantime, especially including any in regard to compliance with stats standards.
