### 2025-04-22 #701 @marcburri
@maurolepore @mpadge Thanks for having a look at our package. I was not sure whether it would be a good fit for a statistical package. But in that case I'll go ahead and insert the `srrstats` tags.
