### 2025-04-23 4d5ed91
Update pkgcheck.yaml
