+++ .github/workflows/pkgcheck.yaml
@@ -21,4 +21,4 @@ jobs:
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
         with:
-          summary-only: false
+          post-to-issue: false
