+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   reviser: reviser.html
   revision-analysis: revision-analysis.html
   understanding-revisions: understanding-revisions.html
-last_built: 2025-04-23T11:22Z
+last_built: 2025-04-23T12:12Z
 urls:
   reference: https://p-wegmueller.github.io/reviser/reference
   article: https://p-wegmueller.github.io/reviser/articles
