+++ R/revisions.R
@@ -42,12 +42,43 @@
 #' Input validation ensures that only one of `ref_date`, `nth_release`, or
 #' `interval` is specified.
 #'
+#' @srrstats {G1.3} Statistical terminology for time series revisions defined
+#' @srrstats {G1.4} Using roxygen2 for function documentation
+#' @srrstats {G2.0} Implements assertions on lengths of inputs through
+#' validation checks
+#' @srrstats {G2.1} Implements assertions on types of inputs (e.g., date,
+#' numeric)
+#' @srrstats {G2.1a} Documents expectations on data types for all inputs
+#' @srrstats {G2.3} Uses validation logic to ensure inputs are exclusive
+#' @srrstats {G2.3a} Uses validation logic to ensure parameters have expected
+#' values
+#' @srrstats {G2.3b} use `tolower()`
+#' @srrstats {G2.4} convert between different data types
+#' @srrstats {G2.4a} Converts input as needed (ref_date converted to Date)
+#' @srrstats {G2.7} Accepts both wide and long format tabular data through
+#' vintages_check
+#' @srrstats {G2.8} Provides conversion routines (vintages_long) to ensure
+#' consistent input format
+#' @srrstats {G5.8} Handles edge conditions with appropriate error messages
+#' @srrstats {TS1.0} Uses explicit class systems for time series data
+#' (tbl_revision)
+#' @srrstats {TS1.1} Explicitly documents required input data structure
+#' @srrstats {TS1.2} Implements validation routines to confirm inputs are
+#' acceptable
+#' @srrstats {TS1.3} Uses a single pre-processing routine (vintages_check,
+#' vintages_long)
+#' @srrstats {TS1.4} Maintains time-based components of input data
+#' @srrstats {TS1.5} Ensures ordering of time through
+#' dplyr::arrange(pub_date, time)
+#' @srrstats {TS4.0a} Returns values in same class as input data
+#' @srrstats {TS4.2} Type and class of return values  documented
+#'
 #' @examples
 #' # Example data
 #' df <- dplyr::filter(reviser::gdp , id=="US")
 #'
 #' # Calculate revisions using an interval of 1
-#' revisions_interval <- get_revisions(df, interval = 1)
+#' revisions_interval <- get_revisions(df, interval = 1L)
 #'
 #' # Calculate revisions using a fixed reference date
 #' revisions_date <- get_revisions(df, ref_date = as.Date("2023-02-01"))
@@ -71,21 +102,33 @@ get_revisions <- function(
 
   # Default interval is 1
   if (specified_count == 0L) {
-    interval <- 1
+    interval <- 1L
   }
 
   if (specified_count > 1L) {
     rlang::abort("Specify only one of 'ref_date', 'nth_release' or 'interval'.")
   }
 
-  if (!is.null(ref_date)) {
+  if (!is.null(interval)) {
+    # check interval is length 1 and integer
+    if (length(interval) != 1) {
+      rlang::abort("'interval' must be of length 1.")
+    }
+
+    if (is.integer(interval)) {
+      # Already valid
+    } else if (is.numeric(interval)) {
+      if (interval %% 1 != 0) {
+        rlang::abort("'interval' must be a whole number.")
+      }
+      interval <- as.integer(interval)
+    } else {
+      rlang::abort("'interval' must be an integer or numeric of length 1.")
+    }
+  } else if (!is.null(ref_date)) {
     ref_date <- tryCatch(as.Date(ref_date), error = function(e) e)
     if (!c("Date") %in% class(ref_date)) {
       rlang::abort("The input 'ref_date' must be a date object.")
-    }
-  } else if (!is.null(interval)) {
-    if (!is.numeric(interval) || interval < 1) {
[…]
