+++ man/get_fixed_release.Rd
@@ -9,8 +9,8 @@ get_fixed_release(df, years, month = NULL, quarter = NULL)
 \arguments{
 \item{df}{A data frame containing columns \code{pub_date} (publication date) and \code{ time} (observation date).}
 
-\item{years}{The number of years after \code{pub_date} for which the values
-should be extracted.}
+\item{years}{The integer number of unrestricted years after \code{pub_date} for
+which the values should be extracted.}
 
 \item{month}{An optional parameter specifying the target month as a name
 ("July") or an integer (7). Cannot be used with \code{quarter}.}

[…]
