### 2025-04-26 38b38b3
adjust srr descriptions
