+++ R/srr-stats-standards.R
@@ -17,165 +17,63 @@ NULL
 #' to `@srrstatsNA`, and placed together in this block, along with explanations
 #' for why each of these standards have been deemed not applicable.
 #' (These comments may also be deleted at any time.)
-#' @srrstatsNA {G1.5} Software should include all code necessary to reproduce
-#' results which form the basis of performance claims made in associated
-#' publications.
-#' @srrstatsNA {G1.6} Software should include code necessary to compare
-#' performance claims with alternative implementations in other R packages.
-#' @srrstatsNA {G2.4d} explicit conversion to factor via `as.factor()`
-#' @srrstatsNA {G2.4e} explicit conversion from factor via `as...()` functions
-#' @srrstatsNA {G2.5} Where inputs are expected to be of `factor` type,
-#' secondary documentation should explicitly state whether these should be
-#' `ordered` or not, and those inputs should provide appropriate error or other
-#' routines to ensure inputs follow these expectations.
-#' @srrstatsNA {G2.11} Software should ensure that `data.frame`-like tabular
-#' objects which have columns which do not themselves have standard class
-#' attributes (typically, `vector`) are appropriately processed, and do not
-#' error without reason. This behaviour should be tested. Again, columns
-#' created by the [`units` package](https://github.com/r-quantities/units/)
-#' provide a good test case.
-#' @srrstatsNA {G2.12} Software should ensure that `data.frame`-like tabular
-#' objects which have list columns should ensure that those columns are
-#' appropriately pre-processed either through being removed, converted to
-#' equivalent vector columns where appropriate, or some other appropriate
-#' treatment such as an informative error. This behaviour should be tested.
-#' @srrstatsNA {G2.14c} replace missing data with appropriately imputed values
-#' @srrstatsNA {G2.16} All functions should also provide options to handle
-#' undefined values (e.g., `NaN`, `Inf` and `-Inf`), including potentially
-#' ignoring or removing such values.
-#' @srrstatsNA {G3.0} Statistical software should never compare floating point
-#'  numbers for equality. All numeric equality comparisons should either ensure
-#'  that they are made between integers, or use appropriate tolerances for
-#'  approximate equality.
-#' @srrstatsNA {G3.1} Statistical software which relies on covariance
-#' calculations should enable users to choose between different algorithms for
-#' calculating covariances, and should not rely solely on covariances from the
-#' `stats::cov` function.
-#' @srrstatsNA {G3.1a} The ability to use arbitrarily specified covariance
-#' methods should be documented (typically in examples or vignettes).
-#' @srrstatsNA {G4.0} Statistical Software which enables outputs to be written
-#' to local files should parse parameters specifying file names to ensure
-#' appropriate file suffices are automatically generated where not provided.
-#' @srrstatsNA {G5.0} Where applicable or practicable, tests should use
-#' standard data sets with known properties (for example, the
-#' [NIST Standard Reference Datasets](https://www.itl.nist.gov/div898/strd/), or
-#'  data sets provided by other widely-used R packages).
-#' @srrstatsNA {G5.1} Data sets created within, and used to test, a package
-#' should be exported (or otherwise made generally available) so that users can
[…]
