+++ R/data-vintages.R
@@ -31,6 +31,8 @@
 #'   * Japan
 #'   * Sources: Cabinet Office (Japan), OECD
 #'
+#' @srrstats {G1.0} academic literatre
+#'
 #' @references Indergand, R., Leist, S. A Real-Time Data Set for Switzerland.
 #' Swiss J Economics Statistics 150, 331–352 (2014).
 #'  \url{https://doi.org/10.1007/BF03399410}
