### 2025-04-26 #701 @marcburri
@maurolepore @mpadge The srr tags are inserted. What are now the next steps in the process?
