+++ R/revisions.R
@@ -817,12 +817,13 @@ summary.lst_efficient <- function(object, ...) {
 #' @examples
 #' # Example usage:
 #' df <- dplyr::select(
+#'   get_nth_release(
 #'     na.omit(
 #'       tsbox::ts_pc(
 #'         reviser::gdp)
 #'             ),
 #'    n = 0:10
-#'  )
+#'  ),
 #' -pub_date
 #' )
 #'
