### 2025-05-04 #701 @maurolepore
This marks the end of my EiC rotation and leaves a note for the next one.

* @mpadge is guiding the process towards a full submission.

@marcburri thanks a lot for sharing your work with rOpenSci!
