### 2025-05-20 #701 @mpadge
@marcburri Apologies for being a bit slow moving on to the next step.

---

Two minor issues arise in the checks above, which should be pretty easy for you to address. Could you please:

- Group your functions in your [pkgdown website](https://p-wegmueller.github.io/reviser/reference/index.html)? Information on how to do that is in [our _Dev Guide_](https://devguide.ropensci.org/pkg_building.html#function-grouping)
- Mention your CONTRIBUTING.md within your README, and link to it from there?

#### Stats standards compliance

I also note that your compliance measures are possibly a bit low, with 52.8% of stats standards complied with, and 47.2% not complied with. While our formal requirement is > 50%, which you package meets, we generally find that putting in extra effort to increase compliance before proceeding to review is helpful both for the review process, and in making your package more robust.

Examples include:

- [G2.11 and G2.12](https://stats-devguide.ropensci.org/standards.html#input-structures) for `data.frame` columns in non-standard formats. List columns in particular are commonly returned from tidy (verse/models) workflows, and can easily lead to unexpected errors. Your package should be able to appropriately handle or reject these via explicit error messages.
- G3.1 may be worth considering? I note that you use the {sandwich} package for your covariance calculations, but these are nevertheless hard-coded. That standard should be complied with by allowing alternative covariance calculations to be used.
- You hard-code quite a lot of test data. It may help to consider compliance with G5.0 or G5.1? That will force you to separate data construction from testing itself, and also (if you opt for G5.1) enable others to readily confirm and build upon your test suite.
- Complying with some additional TS standards may also improve your package. Suggestions include [TS1.7](https://stats-devguide.ropensci.org/standards.html#input-data-structures-and-validation-3) and therefore also [TS4.1](https://stats-devguide.ropensci.org/standards.html#return-results-3), and [TS2.0](https://stats-devguide.ropensci.org/standards.html#missing-data),

That said, it is clear that while your package should definitely be considered a time series package, many of those standards simply do not apply. In short: Please run `srr::srr_report()` locally, just go one more time through the NA standards, and consider whether you might be able to comply with a couple more? Thanks :+1:
