+++ R/data-vintages.R
@@ -31,7 +31,7 @@
 #'   * Japan
 #'   * Sources: Cabinet Office (Japan), OECD
 #'
-#' @srrstats {G1.0} academic literatre
+#' @srrstats {G1.0} academic literature
 #'
 #' @references Indergand, R., Leist, S. A Real-Time Data Set for Switzerland.
 #' Swiss J Economics Statistics 150, 331–352 (2014).
