+++ .github/workflows/pkgdown.yaml
@@ -33,7 +33,8 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown, local::.
+          #extra-packages: any::pkgdown, local::.
+          extra-packages: github::r-lib/pkgdown, local::.
           needs: website
 
       - name: Build site
@@ -47,3 +48,4 @@ jobs:
           clean: false
           branch: gh-pages
           folder: docs
+          
