+++ man/plot_vintages.Rd
@@ -119,4 +119,8 @@ plot_vintages(
 }
 \seealso{
 \code{\link[=theme_reviser]{theme_reviser()}}, \code{\link[=scale_color_reviser]{scale_color_reviser()}}, \code{\link[=scale_fill_reviser]{scale_fill_reviser()}}
+
+Other revision_graphs: 
+\code{\link{theme_reviser}()}
 }
+\concept{revision_graphs}
