+++ man/get_first_efficient_release.Rd
@@ -93,3 +93,9 @@ result$e
 Aruoba, S. Borağan, "Revisions Are Not Well Behaved", Journal of
 Money, Credit and Banking, 40(2-3), 319-340, 2008.
 }
+\seealso{
+Other {revision_analysis}: 
+\code{\link{get_revision_analysis}()},
+\code{\link{summary.lst_efficient}()}
+}
+\concept{{revision_analysis}}
