+++ man/get_first_release.Rd
@@ -38,3 +38,13 @@ df <- dplyr::filter(reviser::gdp, id=="US")
 first_release <- get_first_release(df)
 
 }
+\seealso{
+Other {revision_utilities}: 
+\code{\link{get_days_to_release}()},
+\code{\link{get_fixed_release}()},
+\code{\link{get_latest_release}()},
+\code{\link{get_nth_release}()},
+\code{\link{get_releases_by_date}()},
+\code{\link{get_revisions}()}
+}
+\concept{{revision_utilities}}
