### 2025-05-27 1978082
add grouped reference
