+++ man/get_releases_by_date.Rd
@@ -40,7 +40,7 @@ releases_on_date <- get_releases_by_date(df, date)
 
 }
 \seealso{
-Other {revision_utilities}: 
+Other revision utilities: 
 \code{\link{get_days_to_release}()},
 \code{\link{get_first_release}()},
 \code{\link{get_fixed_release}()},
@@ -48,4 +48,4 @@ Other {revision_utilities}:
 \code{\link{get_nth_release}()},
 \code{\link{get_revisions}()}
 }
-\concept{{revision_utilities}}
+\concept{revision utilities}
