+++ man/get_revision_analysis.Rd
@@ -174,8 +174,8 @@ results <- get_revision_analysis(
 
 }
 \seealso{
-Other {revision_analysis}: 
+Other revision analysis: 
 \code{\link{get_first_efficient_release}()},
 \code{\link{summary.lst_efficient}()}
 }
-\concept{{revision_analysis}}
+\concept{revision analysis}
