+++ man/get_revisions.Rd
@@ -74,7 +74,7 @@ revisions_nth <- get_revisions(df, nth_release = 1)
 
 }
 \seealso{
-Other {revision_utilities}: 
+Other revision utilities: 
 \code{\link{get_days_to_release}()},
 \code{\link{get_first_release}()},
 \code{\link{get_fixed_release}()},
@@ -82,4 +82,4 @@ Other {revision_utilities}:
 \code{\link{get_nth_release}()},
 \code{\link{get_releases_by_date}()}
 }
-\concept{{revision_utilities}}
+\concept{revision utilities}
