+++ man/get_days_to_release.Rd
@@ -37,7 +37,7 @@ df_with_days <- get_days_to_release(df)
 
 }
 \seealso{
-Other {revision_utilities}: 
+Other revision utilities: 
 \code{\link{get_first_release}()},
 \code{\link{get_fixed_release}()},
 \code{\link{get_latest_release}()},
@@ -45,4 +45,4 @@ Other {revision_utilities}:
 \code{\link{get_releases_by_date}()},
 \code{\link{get_revisions}()}
 }
-\concept{{revision_utilities}}
+\concept{revision utilities}
