+++ man/get_fixed_release.Rd
@@ -36,7 +36,7 @@ dta <- get_fixed_release(df, quarter = 3, years = 3)
 
 }
 \seealso{
-Other {revision_utilities}: 
+Other revision utilities: 
 \code{\link{get_days_to_release}()},
 \code{\link{get_first_release}()},
 \code{\link{get_latest_release}()},
@@ -44,4 +44,4 @@ Other {revision_utilities}:
 \code{\link{get_releases_by_date}()},
 \code{\link{get_revisions}()}
 }
-\concept{{revision_utilities}}
+\concept{revision utilities}
