+++ man/plot_vintages.Rd
@@ -120,7 +120,7 @@ plot_vintages(
 \seealso{
 \code{\link[=theme_reviser]{theme_reviser()}}, \code{\link[=scale_color_reviser]{scale_color_reviser()}}, \code{\link[=scale_fill_reviser]{scale_fill_reviser()}}
 
-Other revision_graphs: 
+Other revision graphs: 
 \code{\link{theme_reviser}()}
 }
-\concept{revision_graphs}
+\concept{revision graphs}
