+++ _pkgdown.yml
@@ -9,7 +9,6 @@ template:
       <script defer src="https://cdn.jsdelivr.net/npm/katex@0.16.11/dist/contrib/auto-render.min.js" integrity="sha384-43gviWU0YVjaDtb/GhzOouOXtZMP/7XUzwPTstBeZFe/+rCMvRwr4yROQP43s0Xk" crossorigin="anonymous" onload="renderMathInElement(document.body);"></script>
 
 reference:
-  - title: Package index
   - subtitle: Revision analysis
     desc: >
       Functions for analyzing revisions of time series.
