### 2025-05-27 e4009b0
conform with G2.11 and G2.12
