+++ R/utils.R
@@ -295,6 +295,8 @@ vintages_wide <- function(df, names_from = "pub_date") {
 #' @srrstats {G2.10} Software ensures that extraction or filtering of single
 #' columns from tabular inputs behaves consistently regardless of the class of
 #' tabular data.
+#' @srrstats {G2.11} Handle list columns
+#' @srrstats {G2.12} Handle list columns
 #' @srrstats {G2.13} Implements appropriate checks for missing data in critical
 #' columns like 'time' and 'pub_date'.
 #' @srrstats {G5.2a} unique and descriptive errors.
@@ -333,67 +335,89 @@ vintages_wide <- function(df, names_from = "pub_date") {
 #' @keywords internal
 #' @noRd
 vintages_check <- function(df) {
-  # Check if the object is a data.frame or tibble
+  # Helper: Check for simple (scalar, non-list) columns
+  check_simple_columns <- function(df) {
+    issues <- lapply(names(df), function(col) {
+      column <- df[[col]]
+      
+      # G2.12: Check for list columns
+      if (is.list(column) && !is.data.frame(column)) {
+        return(paste0("Column '", col, "' is a list column."))
+      }
+      
+      # G2.11: Check for non-atomic or multi-valued entries
+      non_atomic <- which(!vapply(column, is.atomic, logical(1)))
+      if (length(non_atomic) > 0) {
+        return(paste0("Column '", col, "' has non-atomic elements."))
+      }
+      
+      not_scalar <- which(vapply(column, function(x) length(x) != 1, logical(1)))
+      if (length(not_scalar) > 0) {
+        return(paste0("Column '", col, "' has elements that are not scalar values."))
+      }
+      
+      return(NULL)
+    })
+    
+    issues <- unlist(issues)
+    if (length(issues) > 0) {
+      rlang::abort(paste("Invalid columns detected:\n", paste(issues, collapse = "\n")))
+    }
+    return(TRUE)
+  }
+  
+  # === Begin Main Checks ===
+  # Ensure it's a data.frame or tibble
   if (!is.data.frame(df)) {
     rlang::abort("The provided object is not a data.frame or tibble.")
   }
-
-  # Check if the object contains the "time" column
+  
+  # Perform column structure check (G2.11 and G2.12)
+  check_simple_columns(df)
+  
+  # Check for required "time" column
   if (!"time" %in% colnames(df)) {
     rlang::abort("The 'time' column is missing in the data.frame.")
   }
-
-  # Check if "time" is in the correct date format
+  
+  # Validate date format for "time"
   if (!all(!is.na(as.Date(df$time, format = "%Y-%m-%d")))) {
-    rlang::abort(
-      "The 'time' column contains values that are not in the '%Y-%m-%d' format."
-    )
+    rlang::abort("The 'time' column contains values that are not in the '%Y-%m-%d' format.")
   }
-
-  # Check for "long format"
+  
+  # Check for long format
   long_format <- all(c("pub_date", "value") %in% colnames(df)) ||
     all(c("pub_date", "values") %in% colnames(df)) ||
     all(c("release", "value") %in% colnames(df)) ||
     all(c("release", "values") %in% colnames(df))
-
+  
   if (long_format) {
     if ("pub_date" %in% colnames(df)) {
-      # Check if "pub_date" is in the correct date format
       if (!all(!is.na(as.Date(df$pub_date, format = "%Y-%m-%d")))) {
-        rlang::abort(
-          "The 'pub_date' column contains values that are not 
-          in '%Y-%m-%d' format."
-        )
+        rlang::abort("The 'pub_date' column contains values that are not in '%Y-%m-%d' format.")
       }
     }
-
-    if ("release" %in% colnames(df)) {
-    }
+    
[…]
