+++ R/utils.R
@@ -27,7 +27,6 @@
 #' `keep_na` parameter
 #' @srrstats {TS1.1} Explicitly documents input data types and classes
 #' @srrstats {TS1.2} Validates inputs via `vintages_check()`
-#' @srrstats {TS2.0} Warns for possibly implicit missings
 #' @srrstats {TS4.2} Explicitly documents return value types and classes
 #'
 #' @examples
@@ -123,8 +122,6 @@ vintages_long <- function(df, names_to = "pub_date", keep_na = FALSE) {
     } else {
       long_df <- long_df %>%
         dplyr::filter(!is.na(value))
-      message("There might be implicit NA values in the 'value' column. ",
-              "Consider setting 'keep_na = TRUE' if you want to retain them.")
     }
     long_df <- vintages_assign_class(long_df)
     return(long_df)
@@ -274,6 +271,7 @@ vintages_wide <- function(df, names_from = "pub_date") {
 #'
 #' @param df A data frame or tibble containing data vintages. The data frame
 #' must have specific columns depending on whether it is in long or wide format.
+#' 
 #'
 #' @return A string indicating the format of the data:
 #' - `"long"` if the data frame is in long format.
@@ -311,6 +309,7 @@ vintages_wide <- function(df, names_from = "pub_date") {
 #' @srrstats {TS1.3}  pre-processing routine to validate input data and
 #' transform it to a uniform type.
 #' @srrstats {TS1.4} date column always there
+#' @srrstats {TS2.0} Checks for possibly implicit missings in time column
 #' @srrstats {TS4.2} type and class of all return values are documented.
 #'
 #' @examples
@@ -337,7 +336,7 @@ vintages_wide <- function(df, names_from = "pub_date") {
 #' vintages_check(wide_data) # Should return "wide"
 #' @keywords internal
 #' @noRd
-vintages_check <- function(df) {
+vintages_check <- function(df, time_col = "time") {
   # Helper: Check for simple (scalar, non-list) columns
   check_simple_columns <- function(df) {
     issues <- lapply(names(df), function(col) {
@@ -382,6 +381,15 @@ vintages_check <- function(df) {
   if (!"time" %in% colnames(df)) {
     rlang::abort("The 'time' column is missing in the data.frame.")
   }
+  
+  # Check for implicit missings in time column
+  implicit_missings <- check_implicit_missing(df, time_col)
+  if (implicit_missings$total_missing > 0) {
+    df <- make_explicit_missing(df, time_col)
+    rlang::warn("The 'time' column contains implicit missing values. I created 
+        a new 'time' column with explicit NA values. Please check your data.")
+  }
+  
   
   # Validate date format for "time"
   if (!all(!is.na(as.Date(df$time, format = "%Y-%m-%d")))) {
@@ -488,4 +496,120 @@ standardize_val_col <- function(df) {
   df %>%
     dplyr::rename(value = dplyr::any_of(c("value", "values"))) %>%
     suppressMessages()
+}
+
+#' Check for implicit missing dates in time series data
+#' @param data data.frame containing time series data
+#' @param time_col character, name of the date column
+#' @param freq character, frequency of the time series data (default is "auto")
+#' @return list with summary of results
+#' @srrstats {TS2.0} Checks for possibly implicit missings in time column
+#'
+#' @keywords internal
+#' @noRd
+check_implicit_missing <- function(data, time_col, freq = "auto") {
+  # Convert date column to Date if it isn't already
+  data[[time_col]] <- as.Date(data[[time_col]])
+  
+  # Auto-detect frequency if not specified
+  if (freq == "auto") {
[…]
